fix admin dashboard and logout admin and manager

+++ BE/BE/Controllers/PaymentsController.cs
@@ -22,7 +22,7 @@ namespace BE.Controllers
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> Get(int id)
+        public async Task<IActionResult> Get(long id)
         {
             var item = await _service.GetByIdAsync(id);
             if (item == null) return NotFound(new { success = false, data = (object?)null, message = "Not found" });
@@ -37,7 +37,7 @@ namespace BE.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Update(int id, [FromBody] Payments model)
+        public async Task<IActionResult> Update(long id, [FromBody] Payments model)
         {
             var updated = await _service.UpdateAsync(id, model);
             if (updated == null) return NotFound(new { success = false, data = (object?)null, message = "Not found" });
@@ -45,7 +45,7 @@ namespace BE.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(int id)
+        public async Task<IActionResult> Delete(long id)
         {
             var deleted = await _service.DeleteAsync(id);
             if (!deleted) return NotFound(new { success = false, message = "Not found" });
