+++ BE/BE/Services/Interfaces/IPaymentsService.cs
@@ -7,9 +7,9 @@ namespace BE.Services.Interfaces
     public interface IPaymentsService
     {
         Task<IEnumerable<Payments>> GetAllAsync();
-        Task<Payments?> GetByIdAsync(int id);
+        Task<Payments?> GetByIdAsync(long id);
         Task<Payments> AddAsync(Payments payment);
-        Task<Payments?> UpdateAsync(int id, Payments payment);
-        Task<bool> DeleteAsync(int id);
+        Task<Payments?> UpdateAsync(long id, Payments payment);
+        Task<bool> DeleteAsync(long id);
     }
 }
